Change report from bytes/s bits/s
Unlike framerate and CPU usage, bitrate may easily overflow in a
regular integer in the near future. To avoid this, this variable was
scaled to be 64 bits.

+++ plugins/gstperf.h
@@ -45,7 +45,7 @@ struct _GstPerf
 
   GstClockTime prev_timestamp;
   guint32 frame_count;
-  guint32 bps;
+  guint64 bps;
   guint32 prev_cpu_total;
   guint32 prev_cpu_idle;
